Add validators to runtime variable editors.
